backend/pyproject.toml
[tool.poetry]
name = "fear-greed-backend"
version = "0.1.0"
description = "Casablanca Fear & Greed Index backend"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
selenium = "^4.20.0"
apscheduler = "^3.10.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
joblib = "^1.3.0"
spacy = "^3.7.0"
openai = "^1.54.0"
redis = "^5.0.0"
hiredis = { version = "^2.2.0", optional = true }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
alembic = "^1.13.0"
prometheus-client = "^0.19.0"
structlog = "^23.2.0"
supabase = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.3.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"



